+++ src/components/mock/filter.js
@@ -1,4 +1,4 @@
-import {FILTER_NAMES} from "../constants.js";
+import {FILTER_NAMES} from "../../constants";
 
 export const generateFilters = () => FILTER_NAMES.map((it) => {
   return {
