+++ src/constants.js
@@ -26,5 +26,4 @@ export const MONTH_NAMES = [
 
 export const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
 
-
 export const ESCAPE = `Escape`;
